Replace np.float, np.in1d removed in NumPy 1.24, NumPy 2.0
  np.float                 -> float            (removed in NumPy 1.24)
  np.in1d                  -> np.isin          (removed in NumPy 2.0)

Each of these raises rather than warning, so execution stops at
the first one on a current environment. The replacements are exact
equivalents - no behavioural or numerical change.

Touches 3 file(s); all still compile.

+++ unicorrn/utils/vision3d/utils/opencv.py
@@ -46,9 +46,9 @@ def registration_with_pnp_ransac(
             [corr_pixels[..., 1], corr_pixels[..., 0]], axis=-1
         )  # (h, w) -> (w, h)
 
-    corr_points = corr_points.astype(np.float)
-    corr_pixels = corr_pixels.astype(np.float)
-    intrinsics = intrinsics.astype(np.float)
+    corr_points = corr_points.astype(float)
+    corr_pixels = corr_pixels.astype(float)
+    intrinsics = intrinsics.astype(float)
 
     _, axis_angle, translation, _ = cv2.solvePnPRansac(
         corr_points,
